ds_api_server/requirements.txt
# gradio==3.26.0
# hupper==1.11
langchain==0.0.166
openai==0.27.2
python-dotenv==1.0.0
google-api-python-client==2.86.0
duckduckgo-search==2.9.5
gTTS
openai-whisper
fastapi==0.65.1
SpeechRecognition
pydub
fastapi_redis_cache
pymongo
uvicorn
google-cloud-texttospeech
python-multipart
aiofiles